+++ internal/clientaccess/client_links_settings_validation.go
@@ -1,10 +1,5 @@
 package clientaccess
 
-import (
-	"errors"
-	"strings"
-)
-
 type ClientLinksSettingsValidation struct{}
 
 func NewClientLinksSettingsValidation() ClientLinksSettingsValidation {
@@ -12,8 +7,5 @@ func NewClientLinksSettingsValidation() ClientLinksSettingsValidation {
 }
 
 func (ClientLinksSettingsValidation) Validate(settings Settings) error {
-	if strings.TrimSpace(settings.Domain) == "" {
-		return errors.New("domain is required to build client links")
-	}
 	return nil
 }